+++ db/schema/index.ts
@@ -6,3 +6,4 @@ export * from "./client";
 export * from "./catalog";
 export * from "./sale";
 export * from "./payment";
+export * from "./cash";